Переименование random-objects,перенос констант в модуль моковых объектов
